+++ StArray.h
@@ -17,8 +17,8 @@ public:
     StArray();
     ~StArray();
     int addStudent(int s_id , char* s_name);
-    int addEE_Course(int s_id, int c_num, char* c_name, int hw_num, int hw_weight);
-    int addCS_Course(int s_id, int c_num, char* c_name, int hw_num, int hw_weight, bool isTkf, char* b_name);
+    int addEE_Course(int s_id, int c_num, char* c_name, int hw_num, double hw_weight);
+    int addCS_Course(int s_id, int c_num, char* c_name, int hw_num, double hw_weight, bool isTkf, char* b_name);
     int setHwGrade(int s_id, int c_num, int hw_num, int hw_grd);
     int setExamGrade(int s_id, int c_num, int exam_grd);
     int setFactor(int c_num, int factor);
